ScarletAI_backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy==3.0.1
SQLAlchemy<2.0
psycopg2-binary==2.9.10
pandas==2.2.3
APScheduler==3.8.1
google-api-python-client==2.19.1
google-auth-httplib2>=0.1.0
google-auth-oauthlib==0.4.6
python-dotenv==1.0.0
PyJWT==2.10.1 
flask-jwt-extended==4.7.1
flask-migrate==4.1.0
scipy==1.15.1
scikit-learn==1.6.1
joblib==1.4.2
dateparser==1.2.0
Flask-SocketIO==5.5.1
Flask-Limiter==3.10.1 
ortools==9.12.4544
marshmallow==3.26.1
cerberus==1.3.7
sendgrid==6.11
twilio=9.5.2
flask_uploads=0.2.1
pyotp=2.9.0